requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pypowerwall>=0.10.0
pyyaml>=6.0
pyarrow>=14.0.0
duckdb>=0.9.0
jinja2>=3.1.0
python-multipart>=0.0.6
